carnet-adresse/Wassaya/backend/requirements.docker.txt
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
python-dotenv==1.2.1
pillow==11.1.0
psycopg2-binary>=2.9
gunicorn>=21.2
